requirements.txt
PySimpleGUI>=4.60.5
openai>=1.40.0
python-dotenv>=1.0.1
